Resolve issue #8, split commands into seperate files
